WhatsappYouTubeSummarizer/requirements.txt
Flask
twilio
youtube-transcript-api
google-generativeai
python-dotenv
